requirements.txt
psycopg2==2.6.2
django-compressor==2.1.1
django-extensions==1.7.6
django-nested-admin==3.0.13
django==1.10.4
django-taggit==0.22.0
django-taggit-live==0.0.7
django-widget-tweaks==1.4.1
python-memcached==1.58
channels==1.0.3
redis==2.10.5
Pillow==4.1.0
